# openspec/specs/quiz/openapi.yaml
openapi: 3.0.3
info:
  title: Quiz Feature API
  description: |
    REST API for the Quiz feature in CourseWise educational platform.
    
    This specification can be viewed using:
    - Swagger Editor: https://editor.swagger.io
    - Postman: Import this file directly
    - Any OpenAPI-compatible tool
    
    **Note**: This is a standalone documentation file. The API does not serve 
    a Swagger UI endpoint - use external tools to view this spec.
  version: 1.0.0
  contact:
    name: Team RNA
    
servers:
  - url: http://localhost:9002
    description: Local development server
  - url: '{baseUrl}'
    description: Production or external service
    variables:
      baseUrl:
        default: http://localhost:9002
        description: Base URL configured via NEXT_PUBLIC_QUIZ_SERVICE_URL

tags:
  - name: Quizzes
    description: Quiz management operations
  - name: Quiz Generation
    description: AI-powered quiz generation
  - name: Attempts
    description: Quiz attempt tracking

paths:
  /api/quizzes:
    get:
      tags:
        - Quizzes
      summary: List quizzes
      description: Get all quizzes with optional filtering by course or teacher
      parameters:
        - name: courseId
          in: query
          description: Filter by course ID
          schema:
            type: string
        - name: teacherId
          in: query
          description: Filter by teacher (creator) ID
          schema:
            type: string
      responses:
        '200':
          description: List of quizzes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Quiz'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
                
    post:
      tags:
        - Quizzes
      summary: Create a quiz
      description: Create a new quiz with questions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuizCreate'
      responses:
        '201':
          description: Quiz created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Quiz'
        '500':
          description: Failed to create quiz
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetailed'

  /api/quizzes/{id}:
    get:
      tags:
        - Quizzes
      summary: Get quiz by ID
      description: Retrieve a specific quiz by its ID
      parameters:
        - name: id
          in: path
          required: true
          description: Quiz ID
          schema:
            type: string
      responses:
        '200':
          description: Quiz found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Quiz'
        '404':
          description: Quiz not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
                
    put:
      tags:
        - Quizzes
      summary: Update a quiz
      description: Update an existing quiz
      parameters:
        - name: id
          in: path
          required: true
          description: Quiz ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuizUpdate'
      responses:
        '200':
          description: Quiz updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Quiz'
        '500':
          description: Failed to update quiz
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
                
    delete:
      tags:
        - Quizzes
      summary: Delete a quiz
      description: Delete a quiz by ID
      parameters:
        - name: id
          in: path
          required: true
          description: Quiz ID
          schema:
            type: string
      responses:
        '204':
          description: Quiz deleted successfully
        '500':
          description: Failed to delete quiz
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/quizzes/generate:
    post:
      tags:
        - Quiz Generation
      summary: Generate quiz with AI
      description: |
        Generate quiz questions from course content using AI (Genkit + Gemini).
        
        **Performance**: Generation typically takes 15-30 seconds for 5-10 questions.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuizGenerateRequest'
      responses:
        '200':
          description: Quiz questions generated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuizGenerateResponse'
        '400':
          description: Missing required fields
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Failed to generate quiz
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetailed'

  /api/attempts:
    get:
      tags:
        - Attempts
      summary: List attempts
      description: Get quiz attempts filtered by quiz ID and/or student ID
      parameters:
        - name: quizId
          in: query
          description: Filter by quiz ID
          schema:
            type: string
        - name: studentId
          in: query
          description: Filter by student ID
          schema:
            type: string
      responses:
        '200':
          description: List of attempts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/QuizAttempt'
        '400':
          description: Missing required query parameter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
                
    post:
      tags:
        - Attempts
      summary: Create an attempt
      description: Start a new quiz attempt
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuizAttemptCreate'
      responses:
        '201':
          description: Attempt created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuizAttempt'
        '500':
          description: Failed to create attempt
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetailed'

  /api/attempts/{id}:
    get:
      tags:
        - Attempts
      summary: Get attempt by ID
      description: Retrieve a specific attempt by its ID
      parameters:
        - name: id
          in: path
          required: true
          description: Attempt ID
          schema:
            type: string
      responses:
        '200':
          description: Attempt found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuizAttempt'
        '404':
          description: Attempt not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
                
    put:
      tags:
        - Attempts
      summary: Update an attempt
      description: Update an existing attempt (e.g., submit answers)
      parameters:
        - name: id
          in: path
          required: true
          description: Attempt ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuizAttemptUpdate'
      responses:
        '200':
          description: Attempt updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuizAttempt'
        '500':
          description: Failed to update attempt
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
                
    delete:
      tags:
        - Attempts
      summary: Delete an attempt
      description: Delete an attempt by ID
      parameters:
        - name: id
          in: path
          required: true
          description: Attempt ID
          schema:
            type: string
      responses:
        '204':
          description: Attempt deleted successfully
        '500':
          description: Failed to delete attempt
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

components:
  schemas:
    Quiz:
      type: object
      required:
        - id
        - courseId
        - title
        - questions
        - createdBy
        - createdAt
        - totalPoints
        - difficulty
      properties:
        id:
          type: string
          description: Unique quiz identifier
          example: quiz-abc123
        courseId:
          type: string
          description: Associated course ID
          example: course-xyz
        title:
          type: string
          description: Quiz title
          example: Python Basics Quiz
        description:
          type: string
          description: Quiz description
          example: Test your knowledge of Python fundamentals
        questions:
          type: array
          items:
            $ref: '#/components/schemas/QuizQuestion'
        createdBy:
          type: string
          description: Teacher user ID who created the quiz
          example: teacher-uid-123
        createdAt:
          type: string
          format: date-time
          description: Creation timestamp (ISO 8601)
          example: '2026-01-21T10:30:00Z'
        totalPoints:
          type: integer
          description: Total points possible
          example: 50
        difficulty:
          type: string
          enum: [easy, medium, hard]
          description: Quiz difficulty level
          example: medium
        topics:
          type: array
          items:
            type: string
          description: Topics covered in the quiz
          example: [variables, loops, functions]
          
    QuizCreate:
      type: object
      required:
        - courseId
        - title
        - questions
        - createdBy
        - totalPoints
        - difficulty
      properties:
        courseId:
          type: string
        title:
          type: string
        description:
          type: string
        questions:
          type: array
          items:
            $ref: '#/components/schemas/QuizQuestion'
        createdBy:
          type: string
        totalPoints:
          type: integer
        difficulty:
          type: string
          enum: [easy, medium, hard]
        topics:
          type: array
          items:
            type: string
            
    QuizUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        questions:
          type: array
          items:
            $ref: '#/components/schemas/QuizQuestion'
        totalPoints:
          type: integer
        difficulty:
          type: string
          enum: [easy, medium, hard]
        topics:
          type: array
          items:
            type: string

    QuizQuestion:
      type: object
      required:
        - id
        - questionText
        - questionType
        - correctAnswer
        - explanation
        - points
        - topic
      properties:
        id:
          type: string
          description: Unique question identifier
          example: q1
        questionText:
          type: string
          description: The question text
          example: What is the output of print(2 + 2)?
        questionType:
          type: string
          enum: [multiple-choice, true-false, short-answer]
          description: Type of question
          example: multiple-choice
        options:
          type: array
          items:
            type: string
          description: Answer options (required for multiple-choice)
          example: ['2', '4', '22', 'Error']
        correctAnswer:
          oneOf:
            - type: string
            - type: array
              items:
                type: string
          description: Correct answer(s)
          example: '4'
        explanation:
          type: string
          description: Explanation for the correct answer
          example: The + operator performs addition on integers
        points:
          type: integer
          description: Points for this question
          example: 5
        topic:
          type: string
          description: Topic this question covers
          example: variables

    QuizGenerateRequest:
      type: object
      required:
        - courseContent
        - learningObjectives
        - numberOfQuestions
        - difficulty
      properties:
        courseContent:
          type: string
          description: Course material text to generate questions from
        learningObjectives:
          type: string
          description: Learning objectives for the quiz
        numberOfQuestions:
          type: integer
          minimum: 1
          maximum: 50
          description: Number of questions to generate
          example: 5
        difficulty:
          type: string
          enum: [easy, medium, hard]
          description: Difficulty level for generated questions
        topics:
          type: array
          items:
            type: string
          description: Specific topics to focus on (optional)

    QuizGenerateResponse:
      type: object
      properties:
        questions:
          type: array
          items:
            $ref: '#/components/schemas/QuizQuestion'

    QuizAttempt:
      type: object
      required:
        - id
        - quizId
        - studentId
        - courseId
        - answers
        - score
        - maxScore
        - startedAt
        - status
      properties:
        id:
          type: string
          description: Unique attempt identifier
          example: attempt-xyz789
        quizId:
          type: string
          description: Quiz ID being attempted
          example: quiz-abc123
        studentId:
          type: string
          description: Student user ID
          example: student-uid-456
        courseId:
          type: string
          description: Course ID
          example: course-xyz
        answers:
          type: array
          items:
            $ref: '#/components/schemas/QuizAnswer'
        score:
          type: integer
          description: Current score
          example: 45
        maxScore:
          type: integer
          description: Maximum possible score
          example: 50
        startedAt:
          type: string
          format: date-time
          description: When the attempt started
          example: '2026-01-21T11:00:00Z'
        completedAt:
          type: string
          format: date-time
          nullable: true
          description: When the attempt was completed (null if in-progress)
          example: '2026-01-21T11:15:00Z'
        status:
          type: string
          enum: [in-progress, completed]
          description: Attempt status
          example: completed

    QuizAttemptCreate:
      type: object
      required:
        - quizId
        - studentId
        - courseId
        - maxScore
        - status
      properties:
        quizId:
          type: string
        studentId:
          type: string
        courseId:
          type: string
        answers:
          type: array
          items:
            $ref: '#/components/schemas/QuizAnswer'
          default: []
        score:
          type: integer
          default: 0
        maxScore:
          type: integer
        status:
          type: string
          enum: [in-progress, completed]

    QuizAttemptUpdate:
      type: object
      properties:
        answers:
          type: array
          items:
            $ref: '#/components/schemas/QuizAnswer'
        score:
          type: integer
        completedAt:
          type: string
          format: date-time
        status:
          type: string
          enum: [in-progress, completed]

    QuizAnswer:
      type: object
      required:
        - questionId
        - studentAnswer
        - isCorrect
        - pointsEarned
      properties:
        questionId:
          type: string
          description: Question ID
          example: q1
        studentAnswer:
          oneOf:
            - type: string
            - type: array
              items:
                type: string
          description: Student's answer
          example: '4'
        isCorrect:
          type: boolean
          description: Whether the answer is correct
          example: true
        pointsEarned:
          type: integer
          description: Points earned for this answer
          example: 5

    Error:
      type: object
      properties:
        error:
          type: string
          description: Error description
          example: Quiz not found

    ErrorDetailed:
      type: object
      properties:
        error:
          type: string
          description: Brief error description
        message:
          type: string
          description: Detailed error message
        code:
          type: string
          description: Error code